+++ prisma/migrations/20250507165228_removed_unique_constraint_on_tomb_number/migration.sql
@@ -1,0 +1,2 @@
+-- DropIndex
+DROP INDEX "equipments_tombNumber_key";